#36 fix: correcao de erros do eslint
