+++ src/components/action-buttons/confirm-edit-button/index.tsx
@@ -5,6 +5,7 @@ import { RiEdit2Fill } from 'react-icons/ri';
 import { ActionButton } from '..';
 import { ActionButtonProps } from '../types';
 import { Button } from '@chakra-ui/react';
+
 type ConfirmEditButtonProps<Data> = ActionButtonProps<Data>;
 
 export function ConfirmEditButton<Data>({
